+++ go.mod
@@ -1,8 +1,3 @@
 module github.com/peerbridge/peerbridge
 
 go 1.13
-
-require (
-	golang.org/x/crypto v0.0.0-20191011191535-87dc89f01550
-	golang.org/x/tools v0.0.0-20200612022331-742c5eb664c2 // indirect
-)
